Disable HAProxy support by default for now
This code does not currently work on WinXP:
"The procedure entry point inet_pton could not be located in the dynamic link library WS2_32.dll"

That could likely be addressed using GetProcAddress(), but for now, I want to get to the bottom of a *different* issue reported on a WinXP system, so need to this disabled.
Define HAPROXY_SUPPORT in your build if you need this feature compiled in right now.