Fix ALSA output on Linux
It seems that the "real" ALSA will actually silently fail if you
try to clear errors when no errors have occured.  The FreeBSD
emulation of ALSA does not have this insane requirement, so this
went unnoticed for the 1.1 release.

I suspect that this actually fixes SF bug 24, because ALSA is
preferred over pulseaudio.