Add iniCutSection and iniRemoveSectionFast
iniCutSection is like iniGetSection except that the section is is also
removed from the ini content list. For performance reasons, the list is not
realloced (maybe saving memory), and for this, I added iniRemoveSectionFast.

For cases where an .ini file may contain many sections and keys, it can be
a huge performance improvement to remove the sections as they are parsed
(i.e. use iniCutSection instead of iniGetSection), thus reducing the number of
operations required for subsequent section/key searches (each call to iniGet*).

This change was required to make the new ctrl/????.ini file loading
(scfglib*.c changes) acceptable on Vertrauen, Windows builds.