Resolve failure to touch files when non-owner on *nix
ftouch(), semfile_signal(), and JS file_utime() would fail to update a file's access/mod times with errno=EPERM if not run as the file's owner. From "man utime":

EPERM times is not NULL, the caller's effective UID does not match the owner of the file, and the caller is not privileged (Linux: does not have the CAP_FOWNER capability).

So use a NULL times parameter value when updating to a file's time stamp(s) to the current time.