Insure all builds of safe_strerror() write the string to the passed buf
The GNU_SOURCE build of this function was (sometimes?) just returning the
string and not actually copying it to the passed buf. This is consistent
with the GNU manpage on strerror_r():
"This may be either a pointer to a string that the function stores in buf ..."
but was inconsistent with all the other build types of this function.

Also updated to use strlcpy and write the problematic error number to the
default string (if unknown).