Clean up and improve 6e516ae
The first entry I tried adding was "Example..." which was not a
valid hostname, but was copied into the address and looked silly.

Now only copies into address if it's a valid, usable hostname.

Also, add a convenience macro to check if a connection type uses
the network or not.

New functions in netwrap:
bool isValidHostname(const char *str);
bool isValidAddressString(const char *str);
bool isResolvableHostname(const char *str);

We ust the last one (which calls the other two) for this feature now.