Add support for the '$' field specifier delimiter thing.
Just in case xp_asprintf() is used for text.dat strings.
The field specifier allows you to change the order of parameters,
which is useful for translating (and useless for anything else).

There must not be any "skipped" parameters though, you must use
all of them between the first one and the last one you use.