Add iniAppendSectionWithNamedStrings()
If you have list of named strings (named_string_t) and want to add them as
a section of key/value pairs to a [section] in .ini formatted string list, this
is the function you want to use. I'm not sure if this is exactly what Deuce had
in mind (for use in SyncTERM), but it wasn't a lot of code to write and might
be useful to someone someday.