CP437 0x7C is UNICODE_BROKEN_BAR (U+00A6)
Contrary to pouplar belief CP437 does not encode US-ASCII.
US-ASCII has UNICODE_VERTICAL_LINE (U+007C) there (which does match
Unicode)

Aren't you glad your C compiler didn't use CP437?
if (x ??!??! y) is so much uglier than if (x || y)

Of course, with C99 (or C90 with the 1995 ammendment), we would have
seen a lot of iso646.h and the use of the or, bitor, and or_eq macros,
and maybe that would have gotten that whole list of 11 macros promoted
to keywords eventually (Likely in C23) since it's hard to eat just one
peanut.

Of course, that's all alternate history... it's unlikely that any
compiler vendor would actually care, and | and ¦ (or | and ³ for those
using CP437) would be "the same".