New SyncTERM feature for 1.8b... encrypted personal list
No longer need your BBS passwords and provate keys in plain text on
your hard drive!

Not for SyncTERM v1.7 because this is a fairly invasive change.

This adds a new ini file thing to read and write encrypted files
with a special header so you can read/write plain-text ones with
it too.

Encrypted INI files cannot use include directives (not sure how you
would create such a thing, but don't).

This is almost certain to fail to build some things because you'll
need to build using WITHOUT_CRYPTLIB for things that need to not
use cryptlib, and that would be any Borland thing (I think?)

So I'll break the pipes now, then come back and fix them later.