Move os_*() functions from genwrap.c to new os_info.c
With the addition of iniReadFile() to genwrap.c, all of the things
that use XpDev by just picking the couple wrapper files it needs
out of the dir and building them can't reall use genwrap.c since
there's a lot more that gets pulled in and tracking all that down
is exhausting.

Instead, move the os_*() functions into a separate thing.